ext_awssso: extract common functions among grant_types
