+++ src/agent.rs
@@ -419,7 +419,7 @@ async fn refresh_token_using_oauth2(
 async fn refresh_token_using_awssso(
     session: &crate::session_manager::Session,
 ) -> Option<crate::token::ServerToken> {
-    match crate::oauth_awssso_device_code::refresh_token(&session.token).await {
+    match crate::ext_awssso::refresh_token(&session.token).await {
         Ok(token) => Some(token),
         Err(e) => {
             tracing::warn!(
